Rename pkg and add more tests
